refactor: reorganize server helpers and assets
