+++ internal/server/doc.go
@@ -1,0 +1,7 @@
+package server
+
+/*
+Package server provides HTTP helpers and middleware used by the forum.
+It keeps the main function lean by collecting related utilities here.
+*/
+
