+++ internal/server/load_templates.go
@@ -1,0 +1,36 @@
+package server
+
+import (
+        "html/template"
+        "path/filepath"
+)
+
+/*
+LoadTemplates walks the directory and parses every template alongside the
+shared layout. It returns a map keyed by filename so handlers can pick the
+right template.
+
+Example:
+
+    tpls, err := server.LoadTemplates("./internal/web/templates")
+*/
+func LoadTemplates(dir string) (map[string]*template.Template, error) {
+        layout := filepath.Join(dir, "layout.html")
+        pages, err := filepath.Glob(filepath.Join(dir, "*.html"))
+        if err != nil {
+                return nil, err
+        }
+        m := make(map[string]*template.Template)
+        for _, page := range pages {
+                if filepath.Base(page) == "layout.html" {
+                        continue
+                }
+                t, err := template.ParseFiles(layout, page)
+                if err != nil {
+                        return nil, err
+                }
+                m[filepath.Base(page)] = t
+        }
+        return m, nil
+}
+
