+++ internal/web/templates/404.html
@@ -1,10 +1,11 @@
-{{define "404.html"}}{{template "layout.html" .}}{{end}}
+{{define "404.html"}}{{template "layout.html" .}}{{end}} {{/* 404 error page */}}
 
 {{define "title"}}Page Not Found{{end}}
 {{define "content"}}
-  <div class="card" style="text-align:center; padding:40px;">
+  <!-- 404 error card -->
+  <div class="card center p-40">
     <h1 class="error-code">404</h1>
-    <p style="color:var(--muted); margin-bottom:24px;">
+    <p class="text-muted mb-24">
       The page you’re looking for doesn’t exist or has been moved.
     </p>
     <a href="/" class="btn primary">Go Home</a>
