+++ internal/web/templates/login.html
@@ -1,17 +1,18 @@
-{{define "login.html"}}{{template "layout.html" .}}{{end}}
+{{define "login.html"}}{{template "layout.html" .}}{{end}} {{/* Login page */}}
 
 {{define "title"}}Login{{end}}
 {{define "content"}}
-<div class="card" style="max-width:480px; margin:0 auto;">
+<div class="card max-w-480">
+  <!-- Login form -->
   <form method="post" action="/login">
     <label>Email</label>
     <input type="email" name="email" required>
     <label>Password</label>
     <input type="password" name="password" required>
     {{if .Error}}
-    <p style="color:#f87171">{{.Error}}</p>
+    <p class="text-error">{{.Error}}</p>
     {{end}}
-    <div style="margin-top:10px"><button class="btn primary" type="submit">Login</button></div>
+    <div class="mt-10"><button class="btn primary" type="submit">Login</button></div>
   </form>
 </div>
 {{end}}
