+++ internal/web/templates/post_new.html
@@ -1,8 +1,9 @@
-{{define "post_new.html"}}{{template "layout.html" .}}{{end}}
+{{define "post_new.html"}}{{template "layout.html" .}}{{end}} {{/* New post page */}}
 
 {{define "title"}}New Post{{end}}
 {{define "content"}}
-<div class="card" style="max-width:720px; margin:0 auto;">
+<div class="card max-w-720">
+  <!-- New post form -->
   <form method="post" action="/post/new">
     <label>Title</label>
     <input name="title" required>
@@ -11,14 +12,14 @@
     <label>Categories (choose or type new)</label>
     <div>
       {{range .Categories}}
-        <label style="display:inline-flex;align-items:center;margin-right:8px;">
+        <label class="inline-label">
           <input type="checkbox" name="category" value="{{.}}"> {{.}}
         </label>
       {{end}}
     </div>
     <small>You can also add new categories by typing them separated by commas:</small>
     <input name="category" placeholder="e.g. Announcements, Go">
-    <div style="margin-top:10px"><button class="btn primary" type="submit">Publish</button></div>
+    <div class="mt-10"><button class="btn primary" type="submit">Publish</button></div>
   </form>
 </div>
 {{end}}
