+++ internal/web/templates/post_show.html
@@ -1,15 +1,16 @@
-{{define "post_show.html"}}{{template "layout.html" .}}{{end}}
+{{define "post_show.html"}}{{template "layout.html" .}}{{end}} {{/* Post detail page */}}
 
 {{define "title"}}{{.Post.Title}}{{end}}
 {{define "content"}}
 <article class="card">
-  <h2 style="margin:0 0 6px">{{.Post.Title}}</h2>
-  <div style="font-size:12px;color:#6b7280">by {{.Post.Username}} • {{.Post.CreatedAt}}</div>
+  <!-- Post content -->
+  <h2 class="post-title">{{.Post.Title}}</h2>
+  <div class="meta">by {{.Post.Username}} • {{.Post.CreatedAt}}</div>
   <p>{{.Post.Body}}</p>
   <div>
     {{range .Post.Categories}}<span class="tag">{{.}}</span>{{end}}
   </div>
-  <div style="margin-top:8px">
+  <div class="mt-8">
     <form class="inline" method="post" action="/like">
       <input type="hidden" name="type" value="post">
       <input type="hidden" name="id" value="{{.Post.ID}}">
@@ -26,10 +27,10 @@
 </article>
 
 <section class="card">
-  <h3 style="margin-top:0">Comments</h3>
+  <h3 class="mt-0">Comments</h3>
   {{range .Comments}}
-    <div class="card" style="margin:8px 0">
-      <div style="font-size:12px;color:#6b7280">{{.Username}} • {{.CreatedAt}}</div>
+    <div class="card my-8">
+      <div class="meta">{{.Username}} • {{.CreatedAt}}</div>
       <p>{{.Body}}</p>
       <div>
         <form class="inline" method="post" action="/like">
@@ -51,11 +52,11 @@
   {{end}}
 
   {{if .LoggedIn}}
-  <form method="post" action="/comment/new" style="margin-top:12px">
+  <form method="post" action="/comment/new" class="mt-12">
     <input type="hidden" name="post_id" value="{{.Post.ID}}">
     <label>Your comment</label>
     <textarea name="body" rows="4" required></textarea>
-    <div style="margin-top:8px"><button class="btn primary" type="submit">Add comment</button></div>
+    <div class="mt-8"><button class="btn primary" type="submit">Add comment</button></div>
   </form>
   {{else}}
     <p><a href="/login">Login</a> to comment.</p>
