+++ internal/web/templates/register.html
@@ -1,8 +1,9 @@
-{{define "register.html"}}{{template "layout.html" .}}{{end}}
+{{define "register.html"}}{{template "layout.html" .}}{{end}} {{/* Registration page */}}
 
 {{define "title"}}Register{{end}}
 {{define "content"}}
-<div class="card" style="max-width:480px; margin:0 auto;">
+<div class="card max-w-480">
+  <!-- Registration form -->
   <form method="post" action="/register">
     <label>Email</label>
     <input type="email" name="email" required>
@@ -11,9 +12,9 @@
     <label>Password</label>
     <input type="password" name="password" required>
     {{if .Error}}
-    <p style="color:#f87171">{{.Error}}</p>
+    <p class="text-error">{{.Error}}</p>
     {{end}}
-    <div style="margin-top:10px"><button class="btn primary" type="submit">Create account</button></div>
+    <div class="mt-10"><button class="btn primary" type="submit">Create account</button></div>
   </form>
 </div>
 {{end}}
